Move nosemgrep to its own file
For maintainability
