+++ semgrep/tests/e2e/snapshots/test_ignores/test_nosem_rule/results.json
