+++ semgrep/tests/e2e/snapshots/test_ignores/test_nosem_rule__invalid_id/error.txt
