+++ semgrep/tests/e2e/snapshots/test_ignores/test_nosem_rule_unicode/results.json
