+++ semgrep/tests/e2e/snapshots/test_ignores/test_regex_rule__nosemgrep/results.json
